Revert error renaming
