+++ src/Smart.Design.Razor/TagHelpers/GridList/GridListRowTagHelper.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Razor.TagHelpers;
 using Smart.Design.Razor.TagHelpers.Base;
 using Smart.Design.Razor.TagHelpers.Constants;
 
-namespace Smart.Design.Razor.TagHelpers.List;
+namespace Smart.Design.Razor.TagHelpers.GridList;
 
 /// <summary>
 /// <see cref="ITagHelper" /> Create the grid data in a table body.
